+++ boot.php
@@ -28,6 +28,6 @@ if (rex::isBackend() && rex::getUser()) {
         rex_extension::register('PAGE_TITLE', array('\Cke5\Handler\Cke5ExtensionHandler', 'addIcon'), rex_extension::EARLY);
         rex_extension::register('PAGES_PREPARED', array('\Cke5\Handler\Cke5ExtensionHandler', 'hiddenMain'), rex_extension::EARLY);
         rex_extension::register('REX_FORM_CONTROL_FIELDS', array('\Cke5\Handler\Cke5ExtensionHandler', 'removeDemoControlFields'), rex_extension::LATE);
-        rex_extension::register(array('REX_FORM_SAVED','REX_FORM_DELETED','CKE5_CLONE_PROFILE','CKE5_DELETE_PROFILE'), array('\Cke5\Handler\Cke5ExtensionHandler', 'createProfiles'));
+        rex_extension::register(array('REX_FORM_SAVED', 'REX_FORM_DELETED', 'CKE5_CLONE_PROFILE', 'CKE5_DELETE_PROFILE'), array('\Cke5\Handler\Cke5ExtensionHandler', 'createProfiles'));
     }
 }