+++ lib/Cke5/Creator/Cke5ProfilesApi.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @author mail[at]doerr-softwaredevelopment[dot]com Joachim Doerr
+ * @package redaxo5
+ * @license MIT
+ */
+
+namespace Cke5\Creator;
+
+
+use Cke5\Handler\Cke5DatabaseHandler;
+
+class Cke5ProfilesApi extends Cke5DatabaseHandler
+{
+
+}